Revert "separate menu items and create login menu item"
This reverts commit 23b1055673abe3351ca2ab82b486b65481064a33.
