--- src/components/Layout/Header/Menu/MenuItem.jsx
@@ -1,27 +1,0 @@
-import React from 'react';
-import Link from 'gatsby-link';
-import cN from 'classnames';
-
-import s from '../style.module.less';
-
-const MenuItemLink = ({ slug, isChild, children }) => (
-  <li className={cN(s.navItem, { [s.navItemChild]: isChild })}>
-    <Link
-      className={s.link}
-      activeClassName={s.linkActive}
-      to={slug === '/' ? '/' : `/${slug}/`}
-    >
-      {children}
-    </Link>
-  </li>
-);
-
-const MenuItemButton = ({ onClick, children }) => (
-  <li className={cN(s.navItem)}>
-    <button onClick={onClick} className={s.link}>
-      {children}
-    </button>
-  </li>
-);
-
-export { MenuItemLink, MenuItemButton };
