+++ src/components/Layout/Header/style.module.less
@@ -178,31 +178,6 @@
   cursor: default;
 }
 
-.loginParent {
-  margin-left: 2.5rem;
-
-  .loginParentTitle {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    .loginParentAvatar {
-      width: 2rem;
-      margin-right: 0.75rem;
-    }
-  }
-
-  button {
-    background: none;
-    color: inherit;
-    border: none;
-    padding: 0;
-    font: inherit;
-    cursor: pointer;
-    outline: inherit;
-  }
-}
-
 .link {
   .menuLink();
